+++ modules/apim-adapter/src/main/java/com/axway/apim/lib/StandardExportCLIOptions.java
@@ -48,9 +48,9 @@ public class StandardExportCLIOptions extends CLIOptions {
 		option.setArgName("my/apis");
 		cliOptions.addOption(option);
 		
-		option = new Option("o", "output", true, "Controls the output format. By default the console is used. CSV is not supported for all entities.");
+		option = new Option("o", "output", true, "Controls the output format. By default the console is used. CSV and DAT is not supported for all entities.");
 		option.setRequired(false);
-		option.setArgName("console|json|csv");
+		option.setArgName("console|json|csv|dat");
 		cliOptions.addOption(option);
 	}