+++ modules/apis/src/test/java/com/axway/apim/test/quota/APIBasicQuotaTestIT.java
@@ -53,10 +53,9 @@ public class APIBasicQuotaTestIT extends TestNGCitrusTestRunner {
 			.extractFromPayload("$.[?(@.path=='${apiPath}')].id", "apiId"));
 		
 		echo("####### Check System-Quotas have been setup as configured #######");
-		echo("####### ############ Sleep 5 seconds ##################### #######");
-		Thread.sleep(5000);
+		echo("####### ############ Sleep 2 seconds ##################### #######");
+		Thread.sleep(2000);
 		http(builder -> builder.client("apiManager").send().get("/quotas/00000000-0000-0000-0000-000000000000").header("Content-Type", "application/json"));
-		Thread.sleep(5000);
 		http(builder -> builder.client("apiManager").receive().response(HttpStatus.OK).messageType(MessageType.JSON)
 			.validate("$.restrictions.[?(@.api=='${apiId}')].type", "throttle")
 			.validate("$.restrictions.[?(@.api=='${apiId}')].method", "*")